criada camada de serviços
