+++ TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.cs
@@ -4,6 +4,7 @@ using ApprovalTests;
 using ApprovalTests.Reporters;
 using TheatricalPlayersRefactoringKata.Enums;
 using TheatricalPlayersRefactoringKata.Models;
+using TheatricalPlayersRefactoringKata.Services;
 using Xunit;
 
 namespace TheatricalPlayersRefactoringKata.Tests;
