+++ TheatricalPlayersRefactoringKata.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheatricalPlayersRefactorin
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheatricalPlayersRefactoringKata.Tests", "TheatricalPlayersRefactoringKata.Tests\TheatricalPlayersRefactoringKata.Tests.csproj", "{2AFBF20D-3D97-46EC-B506-8B7C82598951}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheatricalPlayersRefactoringKata.Console", "TheatricalPlayersRefactoringKata.Console\TheatricalPlayersRefactoringKata.Console.csproj", "{985E46F5-E16D-49ED-A331-6B9976AC1262}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheatricalPlayersRefactoringKata.Console", "TheatricalPlayersRefactoringKata.Console\TheatricalPlayersRefactoringKata.Console.csproj", "{985E46F5-E16D-49ED-A331-6B9976AC1262}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TheatricalPlayersRefactoringKata.Services", "TheatricalPlayersRefactoringKata.Services\TheatricalPlayersRefactoringKata.Services.csproj", "{0476506D-6C91-42D8-923B-04C6B430B7F2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{985E46F5-E16D-49ED-A331-6B9976AC1262}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{985E46F5-E16D-49ED-A331-6B9976AC1262}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{985E46F5-E16D-49ED-A331-6B9976AC1262}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0476506D-6C91-42D8-923B-04C6B430B7F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0476506D-6C91-42D8-923B-04C6B430B7F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0476506D-6C91-42D8-923B-04C6B430B7F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0476506D-6C91-42D8-923B-04C6B430B7F2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
